Add ikvm-base plugin
This will allow us to use the bare minimum,
without having the tasks created, even if we don't
care about them.
