+++ src/main/resources/META-INF/gradle-plugins/ikvm-base.properties
@@ -1,0 +1,1 @@
+implementation-class=com.ullink.IkvmBasePlugin
